Implement first functions, pytest environment and documentation
This commit contains the first implementations of code, however none of
the functions are properly running yet. The idea was to first start on
structures however this requires some work in aerodynamics. I have opted
to go for the weissinger method as done in DSE2021. Furthermore I have
looked into how to do the modal analysis, it is all quite a bit of work.
The building of documentation is also already in progress.