DOC: Finalize propulsion notebook
A lot of work has been done on the propulsion notebook. Unfortunately a
lot of work had to be done on the robustness of the analysis. The values
were very likey to crash due to non convergent. Even more so with the
implemention of an error when too high AoA are encoutered. But now we
get nice surface plot which is great!